OGC Coverages: some renames per PR review
- Rename the util module to request.

- Prefix various non-public implementation methods of CoveragesRequest
  with _.
